requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.4.0
anyio==4.11.0
asttokens==3.0.0
async-timeout==4.0.2
attrs==22.1.0
blinker==1.9.0
Bottleneck==1.3.5
brotlipy==0.7.0
cachetools==6.2.1
certifi==2025.10.5
cffi==1.15.1
charset-normalizer==2.1.1
click==8.3.0
cloudpickle==2.2.1
comm==0.1.4
contourpy==1.0.5
cryptography==39.0.1
cycler==0.12.1
cytoolz==0.12.0
dask==2023.4.1
debugpy==1.6.7
decorator==5.1.1
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.1
filelock==3.20.0
fonttools==4.33.3
frozenlist==1.3.3
fsspec==2025.10.0
gmpy2==2.1.2
google-auth==2.43.0
google-auth-oauthlib==0.4.6
grpcio==1.48.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.2
idna==3.4
imagecodecs==2021.8.26
imageio==2.26.0
importlib-metadata==6.0.0
importlib_resources==6.5.2
ipykernel==6.26.0
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.1.1
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.4
lazy_loader==0.1
locket==1.0.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
monai==1.1.0
mpmath==1.3.0
multidict==6.0.2
munkres==1.1.4
nest-asyncio==1.5.8
networkx==2.8.4
nibabel==5.3.2
numexpr==2.8.4
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
oauthlib==3.2.2
opencv-python==4.12.0.88
packaging==23.0
pandas==1.5.3
parso==0.8.3
partd==1.4.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.0.1
platformdirs==4.1.0
ply==3.11
pooch==1.8.2
prompt_toolkit==3.0.52
protobuf==3.20.3
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.21
Pygments==2.19.2
PyJWT==2.4.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7
pyu2f==0.1.5
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==23.0.0
requests==2.29.0
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.6.2
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
setuptools==66.0.0
shellingham==1.5.4
sip==6.6.2
six==1.16.0
sniffio==1.3.1
stack-data==0.6.2
sympy==1.14.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.4
threadpoolctl==3.6.0
tifffile==2021.11.2
timm==1.0.22
toml==0.10.2
toolz==0.12.0
torch==2.9.0
torchvision==0.24.0
tornado==6.2
tqdm==4.65.0
traitlets==5.14.3
triton==3.5.0
typer-slim==0.20.0
typing_extensions==4.15.0
unicodedata2==14.0.0
urllib3==1.26.15
wcwidth==0.2.14
Werkzeug==3.1.3
wheel==0.38.4
yarl==1.8.1
zipp==3.11.0
